+++ original/check-xml-test-suite.py
@@ -2,7 +2,6 @@
 import sys
 import time
 import os
-import string
 sys.path.insert(0, "python")
 import libxml2
 
@@ -14,8 +13,12 @@ test_error = 0
 #
 # the testsuite description
 #
-CONF="xml-test-suite/xmlconf/xmlconf.xml"
+CONF=os.path.join(os.path.dirname(__file__), "xml-test-suite/xmlconf/xmlconf.xml")
 LOG="check-xml-test-suite.log"
+
+if not os.path.exists(CONF):
+    print("Skipping XML Python suite: %s is missing" % CONF)
+    sys.exit(0)
 
 log = open(LOG, "w")
 
@@ -31,9 +34,9 @@ def errorHandler(ctx, str):
     error_nr = error_nr + 1
     if len(error_msg) < 300:
         if len(error_msg) == 0 or error_msg[-1] == '\n':
-	    error_msg = error_msg + "   >>" + str
-	else:
-	    error_msg = error_msg + str
+            error_msg = error_msg + "   >>" + str
+        else:
+            error_msg = error_msg + str
 
 libxml2.registerErrorHandler(errorHandler, None)
 
@@ -58,12 +61,12 @@ def loadNoentDoc(filename):
     ctxt.replaceEntities(1)
     ctxt.parseDocument()
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     if ctxt.wellFormed() != 1:
         doc.freeDoc()
-	return None
+        return None
     return doc
 
 #
@@ -84,15 +87,15 @@ def testNotWf(filename, id):
     ret = ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     if doc != None:
-	doc.freeDoc()
+        doc.freeDoc()
     if ret == 0 or ctxt.wellFormed() != 0:
-        print "%s: error: Well Formedness error not detected" % (id)
-	log.write("%s: error: Well Formedness error not detected\n" % (id))
-	return 0
+        print("%s: error: Well Formedness error not detected" % (id))
+        log.write("%s: error: Well Formedness error not detected\n" % (id))
+        return 0
     return 1
 
 def testNotWfEnt(filename, id):
@@ -110,15 +113,15 @@ def testNotWfEnt(filename, id):
     ret = ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     if doc != None:
-	doc.freeDoc()
+        doc.freeDoc()
     if ret == 0 or ctxt.wellFormed() != 0:
-        print "%s: error: Well Formedness error not detected" % (id)
-	log.write("%s: error: Well Formedness error not detected\n" % (id))
-	return 0
+        print("%s: error: Well Formedness error not detected" % (id))
+        log.write("%s: error: Well Formedness error not detected\n" % (id))
+        return 0
     return 1
 
 def testNotWfEntDtd(filename, id):
@@ -137,15 +140,15 @@ def testNotWfEntDtd(filename, id):
     ret = ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     if doc != None:
-	doc.freeDoc()
+        doc.freeDoc()
     if ret == 0 or ctxt.wellFormed() != 0:
-        print "%s: error: Well Formedness error not detected" % (id)
-	log.write("%s: error: Well Formedness error not detected\n" % (id))
-	return 0
+        print("%s: error: Well Formedness error not detected" % (id))
+        log.write("%s: error: Well Formedness error not detected\n" % (id))
+        return 0
     return 1
 
 def testWfEntDtd(filename, id):
@@ -164,20 +167,20 @@ def testWfEntDtd(filename, id):
     ret = ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     if doc == None or ret != 0 or ctxt.wellFormed() == 0:
-        print "%s: error: wrongly failed to parse the document" % (id)
-	log.write("%s: error: wrongly failed to parse the document\n" % (id))
-	if doc != None:
-	    doc.freeDoc()
-	return 0
+        print("%s: error: wrongly failed to parse the document" % (id))
+        log.write("%s: error: wrongly failed to parse the document\n" % (id))
+        if doc != None:
+            doc.freeDoc()
+        return 0
     if error_nr != 0:
-        print "%s: warning: WF document generated an error msg" % (id)
-	log.write("%s: error: WF document generated an error msg\n" % (id))
-	doc.freeDoc()
-	return 2
+        print("%s: warning: WF document generated an error msg" % (id))
+        log.write("%s: error: WF document generated an error msg\n" % (id))
+        doc.freeDoc()
+        return 2
     doc.freeDoc()
     return 1
 
@@ -197,19 +200,19 @@ def testError(filename, id):
     ret = ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     if doc != None:
-	doc.freeDoc()
+        doc.freeDoc()
     if ctxt.wellFormed() == 0:
-        print "%s: warning: failed to parse the document but accepted" % (id)
-	log.write("%s: warning: failed to parse the document but accepte\n" % (id))
-	return 2
+        print("%s: warning: failed to parse the document but accepted" % (id))
+        log.write("%s: warning: failed to parse the document but accepte\n" % (id))
+        return 2
     if error_nr != 0:
-        print "%s: warning: WF document generated an error msg" % (id)
-	log.write("%s: error: WF document generated an error msg\n" % (id))
-	return 2
+        print("%s: warning: WF document generated an error msg" % (id))
+        log.write("%s: error: WF document generated an error msg\n" % (id))
+        return 2
     return 1
 
 def testInvalid(filename, id):
@@ -227,24 +230,24 @@ def testInvalid(filename, id):
     ret = ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     valid = ctxt.isValid()
     if doc == None:
-        print "%s: error: wrongly failed to parse the document" % (id)
-	log.write("%s: error: wrongly failed to parse the document\n" % (id))
-	return 0
+        print("%s: error: wrongly failed to parse the document" % (id))
+        log.write("%s: error: wrongly failed to parse the document\n" % (id))
+        return 0
     if valid == 1:
-        print "%s: error: Validity error not detected" % (id)
-	log.write("%s: error: Validity error not detected\n" % (id))
-	doc.freeDoc()
-	return 0
+        print("%s: error: Validity error not detected" % (id))
+        log.write("%s: error: Validity error not detected\n" % (id))
+        doc.freeDoc()
+        return 0
     if error_nr == 0:
-        print "%s: warning: Validity error not reported" % (id)
-	log.write("%s: warning: Validity error not reported\n" % (id))
-	doc.freeDoc()
-	return 2
+        print("%s: warning: Validity error not reported" % (id))
+        log.write("%s: warning: Validity error not reported\n" % (id))
+        doc.freeDoc()
+        return 2
         
     doc.freeDoc()
     return 1
@@ -263,24 +266,24 @@ def testValid(filename, id):
     ctxt.parseDocument()
 
     try:
-	doc = ctxt.doc()
+        doc = ctxt.doc()
     except:
         doc = None
     valid = ctxt.isValid()
     if doc == None:
-        print "%s: error: wrongly failed to parse the document" % (id)
-	log.write("%s: error: wrongly failed to parse the document\n" % (id))
-	return 0
+        print("%s: error: wrongly failed to parse the document" % (id))
+        log.write("%s: error: wrongly failed to parse the document\n" % (id))
+        return 0
     if valid != 1:
-        print "%s: error: Validity check failed" % (id)
-	log.write("%s: error: Validity check failed\n" % (id))
-	doc.freeDoc()
-	return 0
+        print("%s: error: Validity check failed" % (id))
+        log.write("%s: error: Validity check failed\n" % (id))
+        doc.freeDoc()
+        return 0
     if error_nr != 0 or valid != 1:
-        print "%s: warning: valid document reported an error" % (id)
-	log.write("%s: warning: valid document reported an error\n" % (id))
-	doc.freeDoc()
-	return 2
+        print("%s: warning: valid document reported an error" % (id))
+        log.write("%s: warning: valid document reported an error\n" % (id))
+        doc.freeDoc()
+        return 2
     doc.freeDoc()
     return 1
 
@@ -294,20 +297,20 @@ def runTest(test):
     uri = test.prop('URI')
     id = test.prop('ID')
     if uri == None:
-        print "Test without ID:", uri
-	return -1
+        print("Test without ID:", uri)
+        return -1
     if id == None:
-        print "Test without URI:", id
-	return -1
+        print("Test without URI:", id)
+        return -1
     base = test.getBase(None)
     URI = libxml2.buildURI(uri, base)
     if os.access(URI, os.R_OK) == 0:
-        print "Test %s missing: base %s uri %s" % (URI, base, uri)
-	return -1
+        print("Test %s missing: base %s uri %s" % (URI, base, uri))
+        return -1
     type = test.prop('TYPE')
     if type == None:
-        print "Test %s missing TYPE" % (id)
-	return -1
+        print("Test %s missing TYPE" % (id))
+        return -1
 
     extra = None
     if type == "invalid":
@@ -316,94 +319,93 @@ def runTest(test):
         res = testValid(URI, id)
     elif type == "not-wf":
         extra =  test.prop('ENTITIES')
-	# print URI
-	#if extra == None:
-	#    res = testNotWfEntDtd(URI, id)
- 	#elif extra == 'none':
-	#    res = testNotWf(URI, id)
-	#elif extra == 'general':
-	#    res = testNotWfEnt(URI, id)
-	#elif extra == 'both' or extra == 'parameter':
-	res = testNotWfEntDtd(URI, id)
-	#else:
-	#    print "Unknown value %s for an ENTITIES test value" % (extra)
-	#    return -1
+        # print URI
+        #if extra == None:
+        #    res = testNotWfEntDtd(URI, id)
+        #elif extra == 'none':
+        #    res = testNotWf(URI, id)
+        #elif extra == 'general':
+        #    res = testNotWfEnt(URI, id)
+        #elif extra == 'both' or extra == 'parameter':
+        res = testNotWfEntDtd(URI, id)
+        #else:
+        #    print "Unknown value %s for an ENTITIES test value" % (extra)
+        #    return -1
     elif type == "error":
-	res = testError(URI, id)
+        res = testError(URI, id)
     else:
         # TODO skipped for now
-	return -1
+        return -1
 
     test_nr = test_nr + 1
     if res > 0:
-	test_succeed = test_succeed + 1
+        test_succeed = test_succeed + 1
     elif res == 0:
-	test_failed = test_failed + 1
+        test_failed = test_failed + 1
     elif res < 0:
-	test_error = test_error + 1
+        test_error = test_error + 1
 
     # Log the ontext
     if res != 1:
-	log.write("   File: %s\n" % (URI))
-	content = string.strip(test.content)
-	while content[-1] == '\n':
-	    content = content[0:-1]
-	if extra != None:
-	    log.write("   %s:%s:%s\n" % (type, extra, content))
-	else:
-	    log.write("   %s:%s\n\n" % (type, content))
-	if error_msg != '':
-	    log.write("   ----\n%s   ----\n" % (error_msg))
-	    error_msg = ''
-	log.write("\n")
+        log.write("   File: %s\n" % (URI))
+        content = string.strip(test.content)
+        while content[-1] == '\n':
+            content = content[0:-1]
+        if extra != None:
+            log.write("   %s:%s:%s\n" % (type, extra, content))
+        else:
+            log.write("   %s:%s\n\n" % (type, content))
+        if error_msg != '':
+            log.write("   ----\n%s   ----\n" % (error_msg))
+            error_msg = ''
+        log.write("\n")
 
     return 0
-	    
 
 def runTestCases(case):
     profile = case.prop('PROFILE')
     if profile != None and \
-       string.find(profile, "IBM XML Conformance Test Suite - Production") < 0:
-	print "=>", profile
+       "IBM XML Conformance Test Suite - Production" not in profile:
+        print("=>", profile)
     test = case.children
     while test != None:
         if test.name == 'TEST':
-	    runTest(test)
-	if test.name == 'TESTCASES':
-	    runTestCases(test)
+            runTest(test)
+        if test.name == 'TESTCASES':
+            runTestCases(test)
         test = test.next
         
 conf = loadNoentDoc(CONF)
 if conf == None:
-    print "Unable to load %s" % CONF
+    print("Unable to load %s" % CONF)
     sys.exit(1)
 
 testsuite = conf.getRootElement()
 if testsuite.name != 'TESTSUITE':
-    print "Expecting TESTSUITE root element: aborting"
+    print("Expecting TESTSUITE root element: aborting")
     sys.exit(1)
 
 profile = testsuite.prop('PROFILE')
 if profile != None:
-    print profile
+    print(profile)
 
 start = time.time()
 
 case = testsuite.children
 while case != None:
     if case.name == 'TESTCASES':
-	old_test_nr = test_nr
-	old_test_succeed = test_succeed
-	old_test_failed = test_failed
-	old_test_error = test_error
+        old_test_nr = test_nr
+        old_test_succeed = test_succeed
+        old_test_failed = test_failed
+        old_test_error = test_error
         runTestCases(case)
-	print "   Ran %d tests: %d succeeded, %d failed and %d generated an error" % (
-	       test_nr - old_test_nr, test_succeed - old_test_succeed,
-	       test_failed - old_test_failed, test_error - old_test_error)
+        print("   Ran %d tests: %d succeeded, %d failed and %d generated an error" % (
+               test_nr - old_test_nr, test_succeed - old_test_succeed,
+               test_failed - old_test_failed, test_error - old_test_error))
     case = case.next
 
 conf.freeDoc()
 log.close()
 
-print "Ran %d tests: %d succeeded, %d failed and %d generated an error in %.2f s." % (
-      test_nr, test_succeed, test_failed, test_error, time.time() - start)
+print("Ran %d tests: %d succeeded, %d failed and %d generated an error in %.2f s." % (
+      test_nr, test_succeed, test_failed, test_error, time.time() - start))
